Feature/nazirah/addfeature2

See merge request Datakraf/datakraf-hr!10

+++ Modules/Leave/Http/Controllers/ConfigsController.php
@@ -8,6 +8,7 @@ use Illuminate\Routing\Controller;
 use Modules\Site\Entities\Center;
 use DB;
 use Datakraf\Day;
+use Modules\Profile\Entities\PersonalDetail;
 
 class ConfigsController extends Controller
 {
@@ -80,16 +81,25 @@ class ConfigsController extends Controller
     public function destroy($id)
     {
         $center= Center::find($id);
-        $center->holidays()->wherePivot('center_id',$id)->detach();
-        $center->delete();
-        toast('Center deleted successfully', 'success', 'top-right');
+        //check center kt user yg assign in center delete
+        $centerexists = PersonalDetail::where('center_id',$id)->exists();
+        if($centerexists == true){
+            toast('Center deleted unsuccessfully. Please remove this following employee from this cost center before deleting', 'error', 'top-right');
+        }
+        else{
+            $center->holidays()->wherePivot('center_id',$id)->detach();
+            $center->delete();
+            toast('Center deleted successfully', 'success', 'top-right');
+        }
         return redirect()->back();
     }
     //add center
     public function addCenter(Request $request){
+      
         $center = Center::create(
             [
                 'name' => $request->name,
+                'code' => $request->code,
                 'address_one' => $request->address_one,
                 'address_two' => $request->address_two,
                 'postcode' => $request->postcode,
@@ -103,8 +113,27 @@ class ConfigsController extends Controller
                 'status' =>1
             ]
         );
-        toast('Center added successfully', 'success', 'top-right');
-        return redirect()->back();
+        // toast('Center added successfully', 'success', 'top-right');
+        // return redirect()->back();
+        return response()->json($center, 200);
 
+    }
+
+    //get code
+    public function getcode(){
+
+        $codecenter = Center::orderBy('id', 'desc')->first();
+        $sub = substr($codecenter->code,1,6);
+        $codecenter = 'C' . sprintf("%06d", $sub + 1);
+        //check exists ka dak
+        $code=Center::where('code',$codecenter)->exists();
+        if($code==true){
+            $codecenter = 'C' . sprintf("%06d", $sub + 1);
+            return response()->json($codecenter, 200);
+        }
+        else{
+            return response()->json($codecenter, 200);
+        }
+        
     }
 }
